+++ R/data.R
@@ -1,0 +1,17 @@
+#' Datasets for testing
+#'
+#' These datasets are used to test compatibility with foreign (spss_foreign),
+#' or haven_2.0 (x_haven_2.0, x_spss_haven_2.0) packages
+#' @rdname test_datasets
+"x_haven_2.0"
+
+#' @rdname test_datasets
+"x_spss_haven_2.0"
+
+#' @rdname test_datasets
+"spss_file"
+
+#' @rdname test_datasets
+"dta_file"
+
+