+++ R/val_labels.R
@@ -54,6 +54,13 @@ val_labels.data.frame <- function(x, prefixed = FALSE) {
 }
 
 #' @export
+`val_labels<-.factor` <- function(x, value) {
+  if (!is.null(value))
+    stop("Value labels cannot be applied to factors.")
+  x %>% remove_attributes("labels")
+}
+
+#' @export
 `val_labels<-.numeric` <- function(x, value) {
   if (!is.null(value)) {
     x <- labelled(x, value, label = var_label(x))
@@ -153,6 +160,19 @@ val_label.data.frame <- function(x, v, prefixed = FALSE) {
 #' @export
 `val_label<-` <- function(x, v, value) {
   UseMethod("val_label<-")
+}
+
+#' @export
+`val_label<-.default` <- function(x, v, value) {
+  # do nothing
+  x
+}
+
+#' @export
+`val_label<-.factor` <- function(x, v, value) {
+  if (!is.null(value))
+    stop("Value labels cannot be applied to factors.")
+  x
 }
 
 #' @export